+++ src/main/java/com/xy/community/controller/ProfileController.java
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import java.util.Comparator;
 
@@ -33,22 +32,6 @@ public class ProfileController {
             @RequestParam(name = "page",defaultValue = "1") Integer page,
             @RequestParam(name = "size",defaultValue = "5") Integer size
     ) {
-//        User user = new User();
-//        Cookie[] cookies = request.getCookies();
-//        if (cookies != null) {
-//            if (cookies != null) {
-//                for (Cookie cookie : cookies) {
-//                    if (cookie.getName().equals("token")) {
-//                        String token = cookie.getValue();
-//                        user = userMapper.findByToken(token);
-//                        if (user != null) {
-//                            request.getSession().setAttribute("user", user);
-//                        }
-//                        break;
-//                    }
-//                }
-//            }
-//        }
         User user = (User) request.getSession().getAttribute("user");
 
         if (user == null) {
@@ -65,9 +48,9 @@ public class ProfileController {
         }
 
         PaginationDTO pagination =  questionService.profileList(user.getId(),page,size);
-        pagination.getPages().sort(new Comparator<Integer>() {
+        pagination.getPages().sort(new Comparator<Long>() {
             @Override
-            public int compare(Integer o1, Integer o2) {
+            public int compare(Long o1, Long o2) {
                 int i = o1 > o2 ? 1 : -1;
                 return i;
             }
