+++ src/main/java/com/xy/community/controller/PublishController.java
@@ -69,7 +69,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
 
         QuestionDTO questionDTO = questionService.getById(id);
